Move v1 currency.ts to v1/
