+++ src/components/Navbar/EthPrice.tsx
@@ -1,9 +1,9 @@
 import CurrencySymbol from 'components/shared/CurrencySymbol'
 
 import { useEtherPrice } from 'hooks/v1/EtherPrice'
-import { currencyName } from 'utils/currency'
+import { currencyName } from 'utils/v1/currency'
 
-import { CURRENCY_USD } from 'constants/currency'
+import { CURRENCY_USD } from 'constants/v1/currency'
 
 export default function EthPrice() {
   const price = useEtherPrice()
