+++ src/components/shared/CurrencySymbol.tsx
@@ -1,6 +1,6 @@
 import { CurrencyOption } from 'models/currency-option'
 import { CSSProperties } from 'react'
-import { currencyStyle, currencySymbol } from 'utils/currency'
+import { currencyStyle, currencySymbol } from 'utils/v1/currency'
 
 export default function CurrencySymbol({
   currency,
