+++ src/components/shared/PayoutModsList.tsx
@@ -19,7 +19,7 @@ import { useContext, useLayoutEffect, useMemo, useState } from 'react'
 import { formatWad, fromPermyriad, fromWad } from 'utils/formatNumber'
 import { amountSubFee } from 'utils/math'
 
-import { CURRENCY_ETH } from 'constants/currency'
+import { CURRENCY_ETH } from 'constants/v1/currency'
 import ProjectPayoutMods from './formItems/ProjectPayoutMods'
 
 export default function PayoutModsList({
