+++ src/components/shared/formItems/ProjectPayoutMods.tsx
@@ -29,7 +29,7 @@ import {
   fromWad,
 } from 'utils/formatNumber'
 import { amountSubFee } from 'utils/math'
-import { currencyName } from 'utils/currency'
+import { currencyName } from 'utils/v1/currency'
 
 import InputAccessoryButton from 'components/shared/InputAccessoryButton'
 import FormattedNumberInput from 'components/shared/inputs/FormattedNumberInput'
