+++ src/components/v1/V1Project/Pay/index.tsx
@@ -10,13 +10,13 @@ import { parseEther } from 'ethers/lib/utils'
 import { useCurrencyConverter } from 'hooks/v1/CurrencyConverter'
 import { CurrencyOption } from 'models/currency-option'
 import { useContext, useMemo, useState } from 'react'
-import { currencyName } from 'utils/currency'
+import { currencyName } from 'utils/v1/currency'
 import { formatWad, fromWad } from 'utils/formatNumber'
 import { decodeFundingCycleMetadata } from 'utils/fundingCycle'
 
 import { disablePayOverrides } from 'constants/v1/overrides'
 import { readNetwork } from 'constants/networks'
-import { CURRENCY_ETH, CURRENCY_USD } from 'constants/currency'
+import { CURRENCY_ETH, CURRENCY_USD } from 'constants/v1/currency'
 import CurrencySymbol from '../../../shared/CurrencySymbol'
 import PayInputSubText from './PayInputSubText'
 import { V1_PROJECT_IDS } from 'constants/v1/projectIds'
