+++ src/components/v1/V1Project/modals/ConfirmPayOwnerModal/index.tsx
@@ -10,7 +10,7 @@ import { constants } from 'ethers'
 import { useCurrencyConverter } from 'hooks/v1/CurrencyConverter'
 import { usePayProjectTx } from 'hooks/v1/transactor/PayProjectTx'
 import { useContext, useState } from 'react'
-import { currencyName } from 'utils/currency'
+import { currencyName } from 'utils/v1/currency'
 import { formattedNum, formatWad } from 'utils/formatNumber'
 import { weightedRate } from 'utils/math'
 import { tokenSymbolText } from 'utils/tokenSymbolText'
