+++ src/components/v2/V2Create/tabs/FundingTabContent.tsx
@@ -21,7 +21,7 @@ import { PayoutMod } from 'models/mods'
 
 import { useETHPaymentTerminalFee } from 'hooks/v2/contractReader/ETHPaymentTerminalFee'
 
-import { CURRENCY_ETH } from 'constants/currency'
+import { CURRENCY_ETH } from 'constants/v1/currency'
 import { shadowCard } from 'constants/styles/shadowCard'
 const { Option } = Select
