+++ src/constants/v1/currency.ts
